+++ OnlineMovieBooking.Domain/Services/UserService/IUserQueryService.cs
@@ -3,10 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using OnlineMovieBooking.Domain.DTO;
+using OnlineMovieBooking.Domain.Repository;
 
-namespace OnlineMovieBooking.Domain.Services.User
+namespace OnlineMovieBooking.Domain.Services.UserService
 {
     public interface IUserQueryService
     {
+        User Find(int id);
     }
 }
